+++ src/components/message/Messages.jsx
@@ -15,7 +15,7 @@ const Message = ({
   subscribeToMore,
   refetch,
   setIsRefetched,
-  setFriends
+  setReceivingMessage
 }) => {
   const toast = useToast();
 
@@ -90,7 +90,7 @@ const Message = ({
             user_id={userData.user_id}
             friend={friend}
             setIsRefetched={setIsRefetched}
-            setFriends={setFriends}
+            setReceivingMessage={setReceivingMessage}
           />
         </div>
